refactor(desktop): extract VersionsPanel from AppDesktop
AppDesktop had grown past 1200 lines after the right-click revision
menu landed. The Versions sidebar block was the most cohesive chunk
in the file with its own context-menu state machine, picking flow,
show-hidden toggle, and click router — moving it out is a clean cut.

VersionsPanel.vue owns the Versions sidebar end-to-end. Workspace-
only operations (Hide, Unhide, Copy hash, Compare with…, Cancel
pick, Escape-to-cancel) run directly against the workspace.
Operations that need the live buffer flushed first (View this
version, Compare to current) emit up to the host, which keeps
ownership of flushSave. Toast messages route through @error / @info.
Mirrors the LibraryTree.vue interface shape.

formatRevisionTimestamp moves to a free function in
desktop/revision-format.ts so both VersionsPanel and the editor's
revision banner (still in AppDesktop) format dates the same way
without one importing the other.

AppDesktop: 1229 → 979 lines (-20%). All 183 vitest cases pass
unchanged — they target user-visible button text and component
names, which the extraction preserves.
